+++ src/elements/Order.php
@@ -34,6 +34,8 @@ class Order extends Element
 
     public $testMode;
 
+    public $paymentType;
+
     /**
      * @var string Number
      */
@@ -78,6 +80,7 @@ class Order extends Element
     public $addressZip;
     // variants
     public $variants;
+    public $postData;
 
     public $dateCreated;
     public $dateOrdered;
@@ -372,6 +375,8 @@ class Order extends Element
         $record->variants = $this->variants;
         $record->transactionInfo = $this->transactionInfo;
         $record->testMode = $this->testMode;
+        $record->paymentType = $this->paymentType;
+        $record->postData = $this->postData;
         $record->save(false);
 
         parent::afterSave($isNew);
@@ -389,7 +394,7 @@ class Order extends Element
     }
 
     /**
-     * @return \craft\base\ElementInterface|null
+     * @return PaymentForm|null
      */
     public function getPaymentForm()
     {