+++ src/web/assets/StripeElementsAsset.php
@@ -9,29 +9,23 @@
 namespace enupal\stripe\web\assets;
 
 use craft\web\AssetBundle;
-use craft\web\assets\cp\CpAsset;
 
-class StripeButtonAsset extends AssetBundle
+class StripeElementsAsset extends AssetBundle
 {
     public function init()
     {
         // define the path that your publishable resources live
         $this->sourcePath = '@enupal/stripe/resources/';
 
-        // define the dependencies
-        $this->depends = [
-            CpAsset::class,
-        ];
-
-        // define the dependencies
         // define the relative path to CSS/JS files that should be registered with the page
         // when this asset bundle is registered
         $this->css = [
-            'css/stripeforms.css'
+            'stripe/css/enupal-stripe-elements.css'
         ];
 
         $this->js = [
-            'js/enupalstripebutton.js'
+            'vendor/js/jquery.min.js',
+            'stripe/js/enupal-stripe-elements.js'
         ];
 
         parent::init();